+++ walkers/src/download.rs
@@ -1,7 +1,10 @@
-use std::path::PathBuf;
+use std::{path::PathBuf, pin::Pin};
 
 use egui::Context;
-use futures::{SinkExt, StreamExt};
+use futures::{
+    future::{select, select_all, Either},
+    SinkExt, StreamExt,
+};
 use image::ImageError;
 use reqwest::header::USER_AGENT;
 use reqwest_middleware::ClientWithMiddleware;
@@ -34,20 +37,38 @@ enum Error {
     Image(ImageError),
 }
 
+struct Download {
+    tile_id: TileId,
+    result: Result<Texture, Error>,
+}
+
 /// Download and decode the tile.
 async fn download_and_decode(
     client: &ClientWithMiddleware,
-    url: &str,
+    tile_id: TileId,
+    url: String,
+    egui_ctx: &Context,
+) -> Download {
+    log::debug!("Downloading '{}'.", url);
+    Download {
+        tile_id,
+        result: download_and_decode_impl(client, url, egui_ctx).await,
+    }
+}
+
+async fn download_and_decode_impl(
+    client: &ClientWithMiddleware,
+    url: String,
     egui_ctx: &Context,
 ) -> Result<Texture, Error> {
     let image = client
-        .get(url)
+        .get(&url)
         .header(USER_AGENT, "Walkers")
         .send()
         .await
         .map_err(Error::HttpMiddleware)?;
 
-    log::debug!("Downloaded {:?}.", image.status());
+    log::debug!("Downloaded '{}': {:?}.", url, image.status());
 
     let image = image
         .error_for_status()
@@ -59,11 +80,48 @@ async fn download_and_decode(
     Texture::new(&image, egui_ctx).map_err(Error::Image)
 }
 
+async fn download_complete(
+    mut tile_tx: futures::channel::mpsc::Sender<(TileId, Texture)>,
+    egui_ctx: Context,
+    tile_id: TileId,
+    result: Result<Texture, Error>,
+) -> Result<(), ()> {
+    match result {
+        Ok(tile) => {
+            tile_tx.send((tile_id, tile)).await.map_err(|_| ())?;
+            egui_ctx.request_repaint();
+        }
+        Err(e) => {
+            log::warn!("{}", e);
+        }
+    };
+
+    Ok(())
+}
+
+enum Downloads<F> {
+    None,
+    Ongoing(Vec<Pin<Box<F>>>),
+    OngoingSaturated(Vec<Pin<Box<F>>>),
+}
+
+impl<F> Downloads<F> {
+    fn new(downloads: Vec<Pin<Box<F>>>) -> Self {
+        if downloads.is_empty() {
+            Self::None
+        } else if downloads.len() < 6 {
+            Self::Ongoing(downloads)
+        } else {
+            Self::OngoingSaturated(downloads)
+        }
+    }
+}
+
 async fn download_continuously_impl<S>(
     source: S,
     http_options: HttpOptions,
     mut request_rx: futures::channel::mpsc::Receiver<TileId>,
-    mut tile_tx: futures::channel::mpsc::Sender<(TileId, Texture)>,
+    tile_tx: futures::channel::mpsc::Sender<(TileId, Texture)>,
     egui_ctx: Context,
 ) -> Result<(), ()>
 where
@@ -71,20 +129,51 @@ where
 {
     // Keep outside the loop to reuse it as much as possible.
     let client = http_client(http_options);
+    let mut downloads = Downloads::None;
 
     loop {
-        let request = request_rx.next().await.ok_or(())?;
-        let url = source.tile_url(request);
-
-        log::debug!("Getting {:?} from {}.", request, url);
-
-        match download_and_decode(&client, &url, &egui_ctx).await {
-            Ok(tile) => {
-                tile_tx.send((request, tile)).await.map_err(|_| ())?;
-                egui_ctx.request_repaint();
+        downloads = match downloads {
+            Downloads::None => {
+                let request = request_rx.next().await.ok_or(())?;
+                let url = source.tile_url(request);
+                let download = download_and_decode(&client, request, url, &egui_ctx);
+                Downloads::Ongoing(vec![Box::pin(download)])
             }
-            Err(e) => {
-                log::warn!("Could not download '{}': {}", &url, e);
+            Downloads::Ongoing(ref mut downloads) => {
+                let download = select_all(downloads.drain(..));
+                match select(request_rx.next(), download).await {
+                    // New download was requested.
+                    Either::Left((request, downloads)) => {
+                        let request = request.ok_or(())?;
+                        let url = source.tile_url(request);
+                        let download = download_and_decode(&client, request, url, &egui_ctx);
+                        let mut downloads = downloads.into_inner();
+                        downloads.push(Box::pin(download));
+                        Downloads::new(downloads)
+                    }
+                    // Ongoing download was completed.
+                    Either::Right(((result, _, downloads), _)) => {
+                        download_complete(
+                            tile_tx.to_owned(),
+                            egui_ctx.to_owned(),
+                            result.tile_id,
+                            result.result,
+                        )
+                        .await?;
+                        Downloads::new(downloads)
+                    }
+                }
+            }
+            Downloads::OngoingSaturated(ref mut downloads) => {
+                let (result, _, downloads) = select_all(downloads.drain(..)).await;
+                download_complete(
+                    tile_tx.to_owned(),
+                    egui_ctx.to_owned(),
+                    result.tile_id,
+                    result.result,
+                )
+                .await?;
+                Downloads::Ongoing(downloads)
             }
         }
     }